Move filter_iterator.hpp to EVMC include dir
This makes the filter_iterator.hpp a public header of EVMC libraries.
