+++ include/evmc/filter_iterator.hpp
