+++ test/unittests/filter_iterator_test.cpp
@@ -2,7 +2,7 @@
 // Copyright 2022 The EVMC Authors.
 // Licensed under the Apache License, Version 2.0.
 
-#include <tools/evmc/filter_iterator.hpp>
+#include <evmc/filter_iterator.hpp>
 #include <gtest/gtest.h>
 #include <cctype>
